raphael: overlay: Don't pin launcher app in memory
Similar to what we did for the camera app, unpin the launcher app from
memory as well. While the default launcher (Launcher3) isn't# Please
enter the commit message for your changes. Lines starting particularly
big, it doesn't make much sense to pin because the launcher# with '#'
will be ignored, and an empty message aborts the commit. does not
typically load new resources much. Most of its resources should# already
be loaded in memory after it starts, so pinning the APK is redundant.